REFACTOR: Added a file for each window component
