+++ components/customWindows/mySelfTaughtCurriculum.tsx
@@ -1,0 +1,13 @@
+import Window from "../window";
+
+export default function MySelfTaughtCurriculumWindow(): JSX.Element {
+    return (
+        <Window
+            taskbarTitle="My Self-Taught Curriculum"
+            startVisible={false}
+            CSS_PositionUtilityClass="CSS_MySelfTaughtCurriculum_Position"
+        >
+            My Self-Taught Curriculum
+        </Window>
+    )
+}
