+++ pages/index.tsx
@@ -1,7 +1,9 @@
 import Background from "@/components/background";
-import AboutMeWindow from "@/components/customWindow/aboutMe";
+import AboutMeWindow from "@/components/customWindows/aboutMe";
+import MyCertificationsWindow from "@/components/customWindows/myCertifications";
+import MyProjectsWindow from "@/components/customWindows/myProjects";
+import MySelfTaughtCurriculumWindow from "@/components/customWindows/mySelfTaughtCurriculum";
 import Sidebar from "@/components/sidebar";
-import Window from "@/components/window";
 import { mainMenuSidebarData } from "@/data/sideBarData";
 import Head from "next/head";
 import Script from "next/script";
@@ -17,27 +19,9 @@ export default function Home(): JSX.Element {
                 sidebarData={mainMenuSidebarData}
             />
             <AboutMeWindow />
-            <Window
-                taskbarTitle="My Projects"
-                startVisible={false}
-                CSS_PositionUtilityClass="CSS_MyProjects_Position"
-            >
-                My Projects
-            </Window>
-            <Window
-                taskbarTitle="My Certifications"
-                startVisible={false}
-                CSS_PositionUtilityClass="CSS_MyCertifications_Position"
-            >
-                My Certifications
-            </Window>
-            <Window
-                taskbarTitle="My Self-Taught Curriculum"
-                startVisible={false}
-                CSS_PositionUtilityClass="CSS_MySelfTaughtCurriculum_Position"
-            >
-                My Self-Taught Curriculum
-            </Window>
+            <MyProjectsWindow />
+            <MyCertificationsWindow />
+            <MySelfTaughtCurriculumWindow />
             <Script type="text/javascript" src="/static/elements.js" />
             <Script type="text/javascript" src="/static/draggableWindows.js" />
             <Script type="text/javascript" src="/static/windowsIndexHandler.js" />
